bioinformatics-workflows/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
networkx>=3.0
biopython>=1.83
pydantic>=2.5
PyYAML>=6.0
requests>=2.31
